requirements.txt
streamlit
easyocr
openai
pandas
numpy
pillow